роут /api/users/me перемещен в отдельный файл users_me.py с соответствующим префиксом api
